pyproject.toml
[tool.pdm]
[project]
requires-python = ">=3.10,<3.11"
dependencies = [
]

[project.optional-dependencies]
tensorflow-macos = [
    "tensorflow-macos>=2.12.0",
]
tensorflow-cpu = [
    "tensorflow-cpu>=2.12.0",
]
generate-onnx = [
    "tf2onnx>=1.14.0",
]
server = [
    "pillow>=9.5.0",
    "onnxruntime>=1.15.0",
    "onnxruntime-extensions>=0.8.0",
    "fastapi>=0.96.0",
    "uvicorn[standard]>=0.22.0",
    "python-multipart>=0.0.6",
]
